Re-organize slate code
