+++ src/dummy/Slate.hpp
@@ -1,7 +1,7 @@
 #pragma once
-#include "Container.hpp"
-#include "SlateKey.hpp"
-#include "Array.hpp"
+#include "../slate/Container.hpp"
+#include "../slate/SlateKey.hpp"
+#include "../slate/Array.hpp"
 #include <functional>
 
 class Slate : public Container<3>
